scene_context/requirements.txt
torch
torchvision
pillow
opencv-python